TF-IDF
TF-IDF (Term Frequency-Inverse Document Frequency) is a statistical measure used to evaluate the importance of a word in a document relative to a collection of documents (corpus). It is commonly used in information retrieval and text mining to reflect how relevant a word is to a document in a corpus.
I used it here to give a TF-IDF score to each word and to keep only most frequent words for retrieval system.